requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
requests>=2.31.0
PyJWT[crypto]>=2.8.0
pydantic>=2.10.0
pytest>=7.4.3
python-dotenv>=1.0.0
cryptography>=43.0.0
python-multipart>=0.0.6
httpx>=0.27.0
astrapy>=0.7.0

